+++ tempIdeas/untitled/mathOp.h
@@ -8,11 +8,11 @@ namespace mathOp {
 
 using std::map; using std::string; using std::list;
 
-void doMultiplyOnMap(map<string, float>&m, float&toMultiply);
+void doMultiplyOnMap(map<string, double>&m, double&toMultiply);
 
-list<string> chiaviAmenoB(map<string, float>& A, map<string, float>& B);
+list<string> chiaviAmenoB(map<string, double>& A, map<string, double>& B);
 
-list<string> chiaviAeB(map<string, float>& A, map<string, float>& B);
+list<string> chiaviAeB(map<string, double>& A, map<string, double>& B);
 
 }
 #endif // MATHOP_H
